+++ src/samples/OpenTelemetry/OpenTelemetryExample/Application/Middleware/StreamingPerformanceMiddleware.cs
@@ -1,5 +1,4 @@
 using Blazing.Mediator;
-using Blazing.Mediator;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;