+++ src/samples/Streaming.Api/Streaming.Api/Middleware/StreamingLoggingMiddleware.cs
@@ -1,5 +1,4 @@
 using Blazing.Mediator;
-using Blazing.Mediator;
 using System.Text.Json;
 
 namespace Streaming.Api.Middleware;